+++ processor/src/main/java/react4j/processor/MethodChecks.java
@@ -135,7 +135,7 @@ final class MethodChecks
    */
   static void mustBeSubclassCallable( @Nonnull final TypeElement targetType,
                                       @Nonnull final String annotationName,
-                                      @Nonnull final ExecutableElement method )
+                                      @Nonnull final Element method )
     throws ReactProcessorException
   {
     mustNotBeStatic( annotationName, method );
@@ -143,9 +143,19 @@ final class MethodChecks
     mustNotBePackageAccessInDifferentPackage( targetType, annotationName, method );
   }
 
+  static void mustBeStaticallySubclassCallable( @Nonnull final TypeElement targetType,
+                                                @Nonnull final String annotationName,
+                                                @Nonnull final Element method )
+    throws ReactProcessorException
+  {
+    mustBeStatic( annotationName, method );
+    mustNotBePrivate( annotationName, method );
+    mustNotBePackageAccessInDifferentPackage( targetType, annotationName, method );
+  }
+
   static void mustNotBePackageAccessInDifferentPackage( @Nonnull final TypeElement component,
                                                         @Nonnull final String annotationName,
-                                                        @Nonnull final ExecutableElement method )
+                                                        @Nonnull final Element method )
     throws ReactProcessorException
   {
     final Set<Modifier> modifiers = method.getModifiers();